:art: Properly arrange the filter applying logic
